adicionado confirmação para deletar

+++ FRONTEND/the-bear/src/app/components/feedback-adm/feedback-adm.component.ts
@@ -1,10 +1,12 @@
 import { LiveAnnouncer } from '@angular/cdk/a11y';
 import { AfterViewInit, Component } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { MatTableDataSource } from '@angular/material/table';
 import { Router } from '@angular/router';
 import { Feedback } from 'src/app/model/feedback';
 import { FeedbackService } from 'src/app/services/feedback/feedback.service';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-feedback-adm',
@@ -17,7 +19,8 @@ export class FeedbackAdmComponent implements AfterViewInit {
     private _liveAnnouncer: LiveAnnouncer,
     private router: Router,
     private feedbackService: FeedbackService,
-    private snackBar: MatSnackBar
+    private snackBar: MatSnackBar,
+    private dialog: MatDialog
   ) { }
 
   feedBackDataSource = new MatTableDataSource<Feedback>();
@@ -48,23 +51,30 @@ export class FeedbackAdmComponent implements AfterViewInit {
   }
 
   // Método para remover feedback
-  removerFeedback(id: number) {
-    this.feedbackService.removerFeedback(id).subscribe(
-      (success) => {
-        if (success) {
-          // Feedback removido com sucesso
-          // Recarrega a página para refletir as alterações
-          location.reload();
-        } else {
-          // Falha ao remover feedback
-          console.error('Falha ao remover feedback!');
-        }
-      },
-      (error) => {
-        // Erro ao fazer a requisição
-        console.error('Erro ao remover feedback:', error);
+  removerFeedback(id: number): void {
+    const dialogRef = this.dialog.open(ConfirmDialogComponent)
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        // O usuário confirmou a exclusão, então podemos prosseguir com a remoção
+        this.feedbackService.removerFeedback(id).subscribe(
+          (success) => {
+            if (success) {
+              // Feedback removido com sucesso
+              // Recarrega a página para refletir as alterações
+              location.reload();
+            } else {
+              // Falha ao remover feedback
+              console.error('Falha ao remover feedback!');
+            }
+          },
+          (error) => {
+            // Erro ao fazer a requisição
+            console.error('Erro ao remover feedback:', error);
+          }
+        );
       }
-    );
+    });
   }
 
 }
